+++ lib/src/flutter_web_bluetooth_interface.dart
@@ -79,4 +79,110 @@ abstract class FlutterWebBluetoothInterface {
   /// See: [RequestOptionsBuilder]
   ///
   Future<BluetoothDevice> requestDevice(RequestOptionsBuilder options);
+
+  ///
+  /// The [advertisements] stream emits an event with a
+  /// [AdvertisementBluetoothDevice]. This device doesn't have a gatt server and
+  /// can thus not do everything you may want.
+  ///
+  /// This method requests the user to pair to the device.
+  ///
+  /// All this method does is constructs a [RequestOptionsBuilder] using
+  /// information from [device], [requiredServices], and [optionalServices] and
+  /// then calls [requestDevice].
+  ///
+  /// There is no guarantee that the user only sees 1 option in their pair
+  /// dialog and thus there is no guarantee that the user pairs the exact same
+  /// device as the one given.
+  ///
+  /// May throw the same exceptions as [requestDevice].
+  ///
+  /// See: [requestDevice]
+  ///
+  Future<BluetoothDevice> requestAdvertisementDevice(
+    AdvertisementBluetoothDevice device, {
+    List<String> requiredServices = const [],
+    List<String> optionalServices = const [],
+  });
+
+  ///
+  /// Check to see if the current browser has the [requestLEScan] method.
+  ///
+  /// Use this to avoid the [NativeAPINotImplementedError].
+  ///
+  bool get hasRequestLEScan;
+
+  ///
+  /// Request the user to start scanning for Bluetooth LE devices in the
+  /// area. Not every browser supports this method yet so check it using
+  /// [hasRequestLEScan]. However even if the browser supports it, the [Future]
+  /// may never complete on browsers. This has been the case for Chrome on linux
+  /// and windows even with the correct flag enabled. Chrome on Android does
+  /// seem to work. Add a [Future.timeout] to combat this.
+  ///
+  /// The devices found through this are emitted using the [advertisements]
+  /// stream. The devices emitted through this stream aren't [BluetoothDevice]s
+  /// but [AdvertisementBluetoothDevice]s instead as they don't have a
+  /// gatt server.
+  ///
+  /// It will only emit devices that match the [options] so it could happen
+  /// that there are no devices in range while the scan is running.
+  /// See [LEScanOptionsBuilder] for details on the options.
+  ///
+  /// Once a scan is running (and there were no errors) it can be stopped by
+  /// calling [BluetoothLEScan.stop] on the returned object from the [Future].
+  /// If this object doesn't get saved then there is no way to stop the scan,
+  /// it should be able to start multiple scans with different scan options.
+  ///
+  /// - May throw [UserCancelledDialogError] if the user cancelled the dialog.
+  ///
+  /// - May throw [NativeAPINotImplementedError] if the browser/ user agent
+  /// doesn't support this method. This may still be thrown even if
+  /// [hasRequestLEScan] is checked first.
+  ///
+  /// - May throw [StateError] for any state error that the method may throw.
+  ///
+  /// - May throw [PolicyError] if Bluetooth has been disabled by an
+  /// administrator via a policy.
+  ///
+  /// - May throw [PermissionError] if the user has disallowed the permission.
+  ///
+  /// - May throw [BluetoothAdapterNotAvailable] if there is no Bluetooth
+  /// adapter available.
+  ///
+  /// - May throw [BrowserError] for every other browser error.
+  ///
+  Future<BluetoothLEScan> requestLEScan(LEScanOptionsBuilder options);
+
+  ///
+  /// the [advertisements] stream emits [AdvertisementReceivedEvent]s
+  /// for devices found through [requestLEScan].
+  ///
+  /// The device that is in this event is a [AdvertisementBluetoothDevice] this
+  /// bluetooth device lacks a gatt server and can thus not communicate with
+  /// any [BluetoothCharacteristic]s. Use [requestAdvertisementDevice] to get
+  /// a [BluetoothDevice] based on the [AdvertisementBluetoothDevice].
+  ///
+  /// Even if the browser doesn't support [requestLEScan] this stream will not
+  /// throw an [Error]. It will just never emit any events since you can't start
+  /// scan.
+  ///
+  Stream<AdvertisementReceivedEvent<AdvertisementBluetoothDevice>>
+      get advertisements;
+
+  ///
+  /// This is a setting for (new) devices if it should use memory for advertisements.
+  ///
+  /// Not every device sends a completely filled out advertisement packet for
+  /// each advertisements. For example every other packet might have the name
+  /// field missing. If this setting is set to `true` it will use the last received
+  /// event to fill in the missing data on the current new event.
+  ///
+  /// You may want to disable this for certain projects in that case set this
+  /// option to `false`.
+  ///
+  /// It can also be set on a per device level if that is desirable.
+  /// [BluetoothDevice.advertisementsUseMemory].
+  ///
+  bool defaultAdvertisementsMemory = true;
 }